Making Python 3 bindings portable (finally ...)
Updated Model::writeResults() to write to location relative to current working directory rather than always relative to data/output/.
Figured out how to get Python 3 bindings to compile on Windows (turned out to be a missing header, namely iso646.h). The setup.py in pybindings/ now works on both Linux and Windows.

+++ test/assets/dispatchable/combustion/test_Diesel.cpp
@@ -141,7 +141,7 @@ try {
         27.254952
     };
     
-    for (int i = 0; i < exp_load_ratio_vec.size(); i++) {
+    for (size_t i = 0; i < exp_load_ratio_vec.size(); i++) {
         testFloatEquals(
             test_diesel_fuel_lookup.fuel_interp_load_ratio_vec[i],
             exp_load_ratio_vec[i],
@@ -172,7 +172,7 @@ try {
         0, 20, 20, 20, 20, 25, 30, 40, 50, 60, 70, 80, 90, 100, 100, 100
     };
     
-    for (int i = 0; i < requested_production_vec_kW.size(); i++) {
+    for (size_t i = 0; i < requested_production_vec_kW.size(); i++) {
         double requested_production_kW = requested_production_vec_kW[i];
         
         testFloatEquals(
@@ -197,7 +197,7 @@ try {
         0, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1
     };
     
-    for (int i = 0; i < requested_production_vec_kW.size(); i++) {
+    for (size_t i = 0; i < requested_production_vec_kW.size(); i++) {
         double requested_production_kW = requested_production_vec_kW[i];
         
         double production_kW = test_diesel.requestProductionkW(
@@ -278,7 +278,7 @@ try {
         26.4290851680678
     };
     
-    for (int i = 0; i < interp_test_load_ratio_vec.size(); i++) {
+    for (size_t i = 0; i < interp_test_load_ratio_vec.size(); i++) {
         testFloatEquals(
             test_diesel_fuel_lookup._fuelConsumptionLookupL(
                 interp_test_load_ratio_vec[i] * struct_disp.cap_kW,
